+++ TombEngine/Game/Lara/PlayerStateMachine.cpp
@@ -130,7 +130,7 @@ namespace TEN::Entities::Player
 		PlayerBehaviorStateRoutines[LS_GIANT_BUTTON_PUSH] = std::pair(lara_as_controlled, lara_default_col);
 		PlayerBehaviorStateRoutines[LS_TRAPDOOR_FLOOR_OPEN] = std::pair(lara_as_controlled, lara_void_func);
 		PlayerBehaviorStateRoutines[LS_FREEFALL_BIS] = std::pair(lara_as_controlled, lara_void_func);
-		PlayerBehaviorStateRoutines[LS_ROUND_HANDLE] = std::pair(lara_as_controlled_no_look, lara_col_turn_switch);
+		PlayerBehaviorStateRoutines[LS_ROUND_HANDLE] = std::pair(lara_as_controlled_no_look_follow, lara_col_turn_switch);
 		PlayerBehaviorStateRoutines[LS_COGWHEEL] = std::pair(lara_as_controlled_no_look, lara_void_func);
 		PlayerBehaviorStateRoutines[LS_LEVERSWITCH_PUSH] = std::pair(lara_as_controlled, lara_void_func);
 		PlayerBehaviorStateRoutines[LS_HOLE] = std::pair(lara_as_pickup, lara_default_col);