+++ TombEngine/Scripting/Internal/TEN/Flow/Settings/Settings.h
@@ -60,6 +60,7 @@ namespace TEN::Scripting
 
 	struct GraphicsSettings
 	{
+		bool AmbientOcclusion = true;
 		bool Skinning = true;
 
 		static void Register(sol::table& parent);